Refatora estrutura de tarefas: App, TaskInput, TaskList e TaskCounter corrigidos
